adjust the files
In this change, I removed the posts.tsx and incorporated the rendering of the API inside the index.tsx file to avoid rendering the same thing
